# Conflicts:
#	opencell-tests/api/soapui/Full-Ws-CRUD-Test-Entities.xml

+++ opencell-api/src/main/java/org/meveo/api/custom/CustomTableApi.java
@@ -157,7 +157,8 @@ public class CustomTableApi extends BaseApi {
         CustomEntityTemplate cet = customTableService.getCET(customTableCode);
         Map<String, CustomFieldTemplate> cfts = customTableService.validateCfts(cet, false);
         pagingAndFiltering.setFilters(customTableService.convertValue(pagingAndFiltering.getFilters(), cfts.values(), true, null));
-        PaginationConfiguration paginationConfig = toPaginationConfiguration(FIELD_ID, SortOrder.ASCENDING, null, pagingAndFiltering, null);
+        List<String> fields = pagingAndFiltering.getFields()!=null?Arrays.asList(pagingAndFiltering.getFields().split(",")):null;
+		PaginationConfiguration paginationConfig = toPaginationConfiguration(FIELD_ID, SortOrder.ASCENDING, fields, pagingAndFiltering, null);
         Long totalCount = customTableService.count(cet.getDbTablename(), paginationConfig);
         CustomTableDataResponseDto result = new CustomTableDataResponseDto();
         result.setPaging(pagingAndFiltering);
@@ -183,6 +184,12 @@ public class CustomTableApi extends BaseApi {
         if (dto.getValues() == null || dto.getValues().isEmpty()) {
             customTableService.remove(cet.getDbTablename());
         } else {
+            Map<Boolean, List<CustomTableRecordDto>> partitionedById = dto.getValues().stream().collect(Collectors.partitioningBy(x->x.getValues().get(FIELD_ID)!=null));
+            List<CustomTableRecordDto> valuesWithoutIds = partitionedById.get(false);
+            
+            if (!valuesWithoutIds.isEmpty()) {
+                throw new ValidationException(valuesWithoutIds.size() + " record(s) to remove are missing the IDs.");
+            }
             Set<Long> ids = extractIds(dto);
             customTableService.remove(cet.getDbTablename(), ids);
         }
